Backend/requirements.txt
absl-py==1.0.0
backcall==0.2.0
cachetools==4.2.4
certifi==2020.4.5.1
charset-normalizer==2.0.9
cycler==0.11.0
dataclasses
decorator==5.1.0
docopt==0.6.2
entrypoints==0.3
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
idna==3.3
importlib-metadata==4.8.2
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==7.1.0
jupyter-core==4.9.1
simplejson
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.2.1
nest-asyncio==1.5.4
numpy==1.18.4
oauthlib==3.1.1
opencv-python==4.2.0.34
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pipreqs==0.4.11
prompt-toolkit==3.0.23
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.4.1
six==1.16.0
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.46.0
traitlets==4.3.3
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
yarg==0.1.9
zipp==3.6.0
